--- a/pkg/cmd/flags/proctree.go
+++ b/pkg/cmd/flags/proctree.go
@@ -20,7 +20,8 @@ Example:
       both         | process tree is built from both events and signals.
   --proctree process-cache=8192   | will cache up to 8192 processes in the tree (LRU cache).
   --proctree thread-cache=4096    | will cache up to 4096 threads in the tree (LRU cache).
-  --proctree disable-procfs-query | Will disable procfs queries during runtime
+  --proctree disable-procfs       | will disable procfs entirely.
+  --proctree disable-procfs-query | will disable procfs quering during runtime.
 
 Use comma OR use the flag multiple times to choose multiple options:
   --proctree source=A,process-cache=B,thread-cache=C
@@ -66,7 +67,9 @@ func PrepareProcTree(cacheSlice []string) (proctree.ProcTreeConfig, error) {
 				default:
 					return config, fmt.Errorf("unrecognized proctree source option: %v", option)
 				}
-				cacheSet = true // at least the default ones
+				if config.Source != proctree.SourceNone {
+					cacheSet = true // at least the default ones
+				}
 				continue
 			}
 			if strings.HasPrefix(value, "process-cache=") {
@@ -93,7 +96,12 @@ func PrepareProcTree(cacheSlice []string) (proctree.ProcTreeConfig, error) {
 				cacheSet = true
 				continue
 			}
-			if strings.HasPrefix(value, "disable-procfs-query") {
+			if value == "disable-procfs" {
+				config.ProcfsInitialization = false
+				config.ProcfsQuerying = false
+				continue
+			}
+			if value == "disable-procfs-query" {
 				config.ProcfsQuerying = false
 				continue
 			}