--- a/pkg/ebpf/controlplane/processes.go
+++ b/pkg/ebpf/controlplane/processes.go
@@ -126,35 +126,44 @@ func (ctrl *Controller) procTreeForkProcessor(args []trace.Argument) error {
 }
 
 func (ctrl *Controller) procTreeExecProcessor(args []trace.Argument) error {
-	var errs []error
-
 	if ctrl.processTree == nil {
 		return nil // process tree is disabled
 	}
 
 	// Process & Event identification arguments (won't exist for regular events)
 	timestamp, err := parse.ArgVal[uint64](args, "timestamp")
-	errs = append(errs, err)
+	if err != nil {
+		return err
+	}
 	taskHash, _ := parse.ArgVal[uint32](args, "task_hash")
-	errs = append(errs, err)
 	parentHash, _ := parse.ArgVal[uint32](args, "parent_hash")
-	errs = append(errs, err)
 	leaderHash, _ := parse.ArgVal[uint32](args, "leader_hash")
-	errs = append(errs, err)
 
 	// Executable
 	cmdPath, err := parse.ArgVal[string](args, "cmdpath")
-	errs = append(errs, err)
+	if err != nil {
+		return err
+	}
 	pathName, err := parse.ArgVal[string](args, "pathname")
-	errs = append(errs, err)
+	if err != nil {
+		return err
+	}
 	dev, err := parse.ArgVal[uint32](args, "dev")
-	errs = append(errs, err)
+	if err != nil {
+		return err
+	}
 	inode, err := parse.ArgVal[uint64](args, "inode")
-	errs = append(errs, err)
+	if err != nil {
+		return err
+	}
 	ctime, err := parse.ArgVal[uint64](args, "ctime")
-	errs = append(errs, err)
+	if err != nil {
+		return err
+	}
 	inodeMode, err := parse.ArgVal[uint16](args, "inode_mode")
-	errs = append(errs, err)
+	if err != nil {
+		return err
+	}
 
 	// Binary Interpreter (or Loader): might come empty from the kernel
 	interPathName, _ := parse.ArgVal[string](args, "interpreter_pathname")
@@ -164,21 +173,22 @@ func (ctrl *Controller) procTreeExecProcessor(args []trace.Argument) error {
 
 	// Real Interpreter
 	interp, err := parse.ArgVal[string](args, "interp")
-	errs = append(errs, err)
+	if err != nil {
+		return err
+	}
 
 	// Others
 	stdinType, err := parse.ArgVal[uint16](args, "stdin_type")
-	errs = append(errs, err)
+	if err != nil {
+		return err
+	}
 	stdinPath, err := parse.ArgVal[string](args, "stdin_path")
-	errs = append(errs, err)
+	if err != nil {
+		return err
+	}
 	invokedFromKernel, err := parse.ArgVal[int32](args, "invoked_from_kernel")
-	errs = append(errs, err)
-
-	// Handle errors
-	for _, err := range errs {
-		if err != nil {
-			return err
-		}
+	if err != nil {
+		return err
 	}
 
 	return ctrl.processTree.FeedFromExec(