--- a/pkg/ebpf/controlplane/processes.go
+++ b/pkg/ebpf/controlplane/processes.go
@@ -2,7 +2,6 @@ package controlplane
 
 import (
 	"github.com/aquasecurity/tracee/pkg/events/parse"
-	"github.com/aquasecurity/tracee/pkg/proctree"
 	"github.com/aquasecurity/tracee/pkg/time"
 	"github.com/aquasecurity/tracee/pkg/utils"
 	"github.com/aquasecurity/tracee/types/trace"
@@ -17,112 +16,95 @@ func (ctrl *Controller) procTreeForkProcessor(args []trace.Argument) error {
 		return nil // process tree is disabled
 	}
 
+	var err error
+	// NOTE: override all the fields of the forkFeed, to avoid any previous data.
+	forkFeed := ctrl.processTree.GetForkFeedFromPool()
+	defer ctrl.processTree.PutForkFeedInPool(forkFeed)
+
 	// NOTE: The "parent" related arguments can be ignored for process tree purposes.
 
 	// Process & Event identification arguments
-	timestamp, err := parse.ArgVal[uint64](args, "timestamp")
+	forkFeed.TimeStamp, err = parse.ArgVal[uint64](args, "timestamp")
 	if err != nil {
 		return err
 	}
 
 	// Parent Process (Go up in hierarchy until parent is a process and not a lwp)
-	parentTid, err := parse.ArgVal[int32](args, "parent_process_tid")
+	forkFeed.ParentTid, err = parse.ArgVal[int32](args, "parent_process_tid")
 	if err != nil {
 		return err
 	}
-	parentNsTid, err := parse.ArgVal[int32](args, "parent_process_ns_tid")
+	forkFeed.ParentNsTid, err = parse.ArgVal[int32](args, "parent_process_ns_tid")
 	if err != nil {
 		return err
 	}
-	parentPid, err := parse.ArgVal[int32](args, "parent_process_pid")
+	forkFeed.ParentPid, err = parse.ArgVal[int32](args, "parent_process_pid")
 	if err != nil {
 		return err
 	}
-	parentNsPid, err := parse.ArgVal[int32](args, "parent_process_ns_pid")
+	forkFeed.ParentNsPid, err = parse.ArgVal[int32](args, "parent_process_ns_pid")
 	if err != nil {
 		return err
 	}
-	parentStartTime, err := parse.ArgVal[uint64](args, "parent_process_start_time")
+	forkFeed.ParentStartTime, err = parse.ArgVal[uint64](args, "parent_process_start_time")
 	if err != nil {
 		return err
 	}
 
 	// Thread Group Leader (might be the same as the "child", if "child" is a process)
-	leaderTid, err := parse.ArgVal[int32](args, "leader_tid")
+	forkFeed.LeaderTid, err = parse.ArgVal[int32](args, "leader_tid")
 	if err != nil {
 		return err
 	}
-	leaderNsTid, err := parse.ArgVal[int32](args, "leader_ns_tid")
+	forkFeed.LeaderNsTid, err = parse.ArgVal[int32](args, "leader_ns_tid")
 	if err != nil {
 		return err
 	}
-	leaderPid, err := parse.ArgVal[int32](args, "leader_pid")
+	forkFeed.LeaderPid, err = parse.ArgVal[int32](args, "leader_pid")
 	if err != nil {
 		return err
 	}
-	leaderNsPid, err := parse.ArgVal[int32](args, "leader_ns_pid")
+	forkFeed.LeaderNsPid, err = parse.ArgVal[int32](args, "leader_ns_pid")
 	if err != nil {
 		return err
 	}
-	leaderStartTime, err := parse.ArgVal[uint64](args, "leader_start_time")
+	forkFeed.LeaderStartTime, err = parse.ArgVal[uint64](args, "leader_start_time")
 	if err != nil {
 		return err
 	}
 
 	// Child (might be a process or a thread)
-	childTid, err := parse.ArgVal[int32](args, "child_tid")
+	forkFeed.ChildTid, err = parse.ArgVal[int32](args, "child_tid")
 	if err != nil {
 		return err
 	}
-	childNsTid, err := parse.ArgVal[int32](args, "child_ns_tid")
+	forkFeed.ChildNsTid, err = parse.ArgVal[int32](args, "child_ns_tid")
 	if err != nil {
 		return err
 	}
-	childPid, err := parse.ArgVal[int32](args, "child_pid")
+	forkFeed.ChildPid, err = parse.ArgVal[int32](args, "child_pid")
 	if err != nil {
 		return err
 	}
-	childNsPid, err := parse.ArgVal[int32](args, "child_ns_pid")
+	forkFeed.ChildNsPid, err = parse.ArgVal[int32](args, "child_ns_pid")
 	if err != nil {
 		return err
 	}
-	childStartTime, err := parse.ArgVal[uint64](args, "start_time") // child_start_time
+	forkFeed.ChildStartTime, err = parse.ArgVal[uint64](args, "start_time") // child_start_time
 	if err != nil {
 		return err
 	}
 
-	timestamp = time.BootToEpochNS(timestamp)
-	childStartTime = time.BootToEpochNS(childStartTime)
-	parentStartTime = time.BootToEpochNS(parentStartTime)
-	leaderStartTime = time.BootToEpochNS(leaderStartTime)
+	forkFeed.TimeStamp = time.BootToEpochNS(forkFeed.TimeStamp)
+	forkFeed.ChildStartTime = time.BootToEpochNS(forkFeed.ChildStartTime)
+	forkFeed.ParentStartTime = time.BootToEpochNS(forkFeed.ParentStartTime)
+	forkFeed.LeaderStartTime = time.BootToEpochNS(forkFeed.LeaderStartTime)
 
-	childHash := utils.HashTaskID(uint32(childTid), childStartTime)
-	parentHash := utils.HashTaskID(uint32(parentTid), parentStartTime)
-	leaderHash := utils.HashTaskID(uint32(leaderTid), leaderStartTime)
+	forkFeed.ChildHash = utils.HashTaskID(uint32(forkFeed.ChildTid), forkFeed.ChildStartTime)
+	forkFeed.ParentHash = utils.HashTaskID(uint32(forkFeed.ParentTid), forkFeed.ParentStartTime)
+	forkFeed.LeaderHash = utils.HashTaskID(uint32(forkFeed.LeaderTid), forkFeed.LeaderStartTime)
 
-	return ctrl.processTree.FeedFromFork(
-		proctree.ForkFeed{
-			TimeStamp:       timestamp,
-			ChildHash:       childHash,
-			ParentHash:      parentHash,
-			LeaderHash:      leaderHash,
-			ParentTid:       parentTid,
-			ParentNsTid:     parentNsTid,
-			ParentPid:       parentPid,
-			ParentNsPid:     parentNsPid,
-			ParentStartTime: parentStartTime,
-			LeaderTid:       leaderTid,
-			LeaderNsTid:     leaderNsTid,
-			LeaderPid:       leaderPid,
-			LeaderNsPid:     leaderNsPid,
-			LeaderStartTime: leaderStartTime,
-			ChildTid:        childTid,
-			ChildNsTid:      childNsTid,
-			ChildPid:        childPid,
-			ChildNsPid:      childNsPid,
-			ChildStartTime:  childStartTime,
-		},
-	)
+	return ctrl.processTree.FeedFromFork(forkFeed)
 }
 
 func (ctrl *Controller) procTreeExecProcessor(args []trace.Argument) error {
@@ -130,89 +112,74 @@ func (ctrl *Controller) procTreeExecProcessor(args []trace.Argument) error {
 		return nil // process tree is disabled
 	}
 
+	var err error
+
+	// NOTE: override all the fields of the execFeed, to avoid any previous data.
+	execFeed := ctrl.processTree.GetExecFeedFromPool()
+	defer ctrl.processTree.PutExecFeedInPool(execFeed)
+
 	// Process & Event identification arguments (won't exist for regular events)
-	timestamp, err := parse.ArgVal[uint64](args, "timestamp")
+	execFeed.TimeStamp, err = parse.ArgVal[uint64](args, "timestamp")
 	if err != nil {
 		return err
 	}
-	taskHash, _ := parse.ArgVal[uint32](args, "task_hash")
-	parentHash, _ := parse.ArgVal[uint32](args, "parent_hash")
-	leaderHash, _ := parse.ArgVal[uint32](args, "leader_hash")
+	execFeed.TaskHash, _ = parse.ArgVal[uint32](args, "task_hash")
+	execFeed.ParentHash, _ = parse.ArgVal[uint32](args, "parent_hash")
+	execFeed.LeaderHash, _ = parse.ArgVal[uint32](args, "leader_hash")
 
 	// Executable
-	cmdPath, err := parse.ArgVal[string](args, "cmdpath")
+	execFeed.CmdPath, err = parse.ArgVal[string](args, "cmdpath")
 	if err != nil {
 		return err
 	}
-	pathName, err := parse.ArgVal[string](args, "pathname")
+	execFeed.PathName, err = parse.ArgVal[string](args, "pathname")
 	if err != nil {
 		return err
 	}
-	dev, err := parse.ArgVal[uint32](args, "dev")
+	execFeed.Dev, err = parse.ArgVal[uint32](args, "dev")
 	if err != nil {
 		return err
 	}
-	inode, err := parse.ArgVal[uint64](args, "inode")
+	execFeed.Inode, err = parse.ArgVal[uint64](args, "inode")
 	if err != nil {
 		return err
 	}
-	ctime, err := parse.ArgVal[uint64](args, "ctime")
+	execFeed.Ctime, err = parse.ArgVal[uint64](args, "ctime")
 	if err != nil {
 		return err
 	}
-	inodeMode, err := parse.ArgVal[uint16](args, "inode_mode")
+	execFeed.InodeMode, err = parse.ArgVal[uint16](args, "inode_mode")
 	if err != nil {
 		return err
 	}
 
 	// // Binary Interpreter (or Loader): might come empty from the kernel
-	// interPathName, _ := parse.ArgVal[string](args, "interpreter_pathname")
-	// interDev, _ := parse.ArgVal[uint32](args, "interpreter_dev")
-	// interInode, _ := parse.ArgVal[uint64](args, "interpreter_inode")
-	// interCtime, _ := parse.ArgVal[uint64](args, "interpreter_ctime")
+	// InterpreterPath, _ := parse.ArgVal[string](args, "interpreter_pathname")
+	// InterpreterDev, _ := parse.ArgVal[uint32](args, "interpreter_dev")
+	// InterpreterInode, _ := parse.ArgVal[uint64](args, "interpreter_inode")
+	// InterpreterCtime, _ := parse.ArgVal[uint64](args, "interpreter_ctime")
 
 	// Real Interpreter
-	interp, err := parse.ArgVal[string](args, "interp")
+	execFeed.Interp, err = parse.ArgVal[string](args, "interp")
 	if err != nil {
 		return err
 	}
 
 	// Others
-	stdinType, err := parse.ArgVal[uint16](args, "stdin_type")
+	execFeed.StdinType, err = parse.ArgVal[uint16](args, "stdin_type")
 	if err != nil {
 		return err
 	}
-	stdinPath, err := parse.ArgVal[string](args, "stdin_path")
+	execFeed.StdinPath, err = parse.ArgVal[string](args, "stdin_path")
 	if err != nil {
 		return err
 	}
-	invokedFromKernel, err := parse.ArgVal[int32](args, "invoked_from_kernel")
+	execFeed.InvokedFromKernel, err = parse.ArgVal[int32](args, "invoked_from_kernel")
 	if err != nil {
 		return err
 	}
 
-	return ctrl.processTree.FeedFromExec(
-		proctree.ExecFeed{
-			TimeStamp:  time.BootToEpochNS(timestamp),
-			TaskHash:   taskHash,
-			ParentHash: parentHash,
-			LeaderHash: leaderHash,
-			CmdPath:    cmdPath,
-			PathName:   pathName,
-			Dev:        dev,
-			Inode:      inode,
-			Ctime:      ctime,
-			InodeMode:  inodeMode,
-			// InterpreterPath:   interPathName,
-			// InterpreterDev:    interDev,
-			// InterpreterInode:  interInode,
-			// InterpreterCtime:  interCtime,
-			Interp:            interp,
-			StdinType:         stdinType,
-			StdinPath:         stdinPath,
-			InvokedFromKernel: invokedFromKernel,
-		},
-	)
+	return ctrl.processTree.FeedFromExec(execFeed)
 }
 
 func (ctrl *Controller) procTreeExitProcessor(args []trace.Argument) error {
@@ -220,42 +187,42 @@ func (ctrl *Controller) procTreeExitProcessor(args []trace.Argument) error {
 		return nil // process tree is disabled
 	}
 
+	var err error
+
+	// NOTE: override all the fields of the exitFeed, to avoid any previous data.
+	exitFeed := ctrl.processTree.GetExitFeedFromPool()
+	defer ctrl.processTree.PutExitFeedInPool(exitFeed)
+
 	// Process & Event identification arguments (won't exist for regular events)
-	timestamp, err := parse.ArgVal[uint64](args, "timestamp")
+	exitFeed.TimeStamp, err = parse.ArgVal[uint64](args, "timestamp")
 	if err != nil {
 		return err
 	}
-	taskHash, err := parse.ArgVal[uint32](args, "task_hash")
+	// time of exit is already a timestamp)
+	exitFeed.TimeStamp = time.BootToEpochNS(exitFeed.TimeStamp)
+
+	exitFeed.TaskHash, err = parse.ArgVal[uint32](args, "task_hash")
 	if err != nil {
 		return err
 	}
-	parentHash, err := parse.ArgVal[uint32](args, "parent_hash")
+	exitFeed.ParentHash, err = parse.ArgVal[uint32](args, "parent_hash")
 	if err != nil {
 		return err
 	}
-	leaderHash, err := parse.ArgVal[uint32](args, "leader_hash")
+	exitFeed.LeaderHash, err = parse.ArgVal[uint32](args, "leader_hash")
 	if err != nil {
 		return err
 	}
 
 	// Exit logic arguments
-	exitCode, err := parse.ArgVal[int64](args, "exit_code")
+	exitFeed.ExitCode, err = parse.ArgVal[int64](args, "exit_code")
 	if err != nil {
 		return err
 	}
-	groupExit, err := parse.ArgVal[bool](args, "process_group_exit")
+	exitFeed.Group, err = parse.ArgVal[bool](args, "process_group_exit")
 	if err != nil {
 		return err
 	}
 
-	return ctrl.processTree.FeedFromExit(
-		proctree.ExitFeed{
-			TimeStamp:  time.BootToEpochNS(timestamp), // time of exit is already a times)p
-			TaskHash:   taskHash,
-			ParentHash: parentHash,
-			LeaderHash: leaderHash,
-			ExitCode:   exitCode,
-			Group:      groupExit,
-		},
-	)
+	return ctrl.processTree.FeedFromExit(exitFeed)
 }