@rscampos 2025-01-27
After running the following line, the output seems not to be aligned with the docs:
```
./dist/tracee --proctree help
Error: Select different options for the process tree.

Example:
  --proctree source=[none|events|signals|both]
      none         | process tree is disabled (default).
      events       | process tree is built from events.
      signals      | process tree is built from signals.
      both         | process tree is built from both events and signals.
  --proctree process-cache=8192   | will cache up to 8192 processes in the tree (LRU cache).
  --proctree thread-cache=4096    | will cache up to 4096 threads in the tree (LRU cache).
  --proctree disable-procfs       | will disable procfs entirely.
  --proctree disable-procfs-query | will disable procfs quering during runtime.

Use comma OR use the flag multiple times to choose multiple options:
  --proctree source=A,process-cache=B,thread-cache=C
  --proctree process-cache=X --proctree thread-cache=Y

```

Docs:
https://github.com/aquasecurity/tracee/blob/c1dd27e335a07327c10f6605d4d59177f6b250cd/docs/docs/advanced/data-sources/builtin/process-tree.md?plain=1#L37-L39

Also, when used `help`, we get a error (`Error: Select different options for the process tree.`).