@geyslan 2025-01-27
@rscampos proctree still uses an old format of help (tracee-ebpf). The majority was migrated to `tracee man` style. For now, I'm leaving it as it's.