@rscampos 2025-01-28
@geyslan I've doubled-checked the performance. Used a AWS t4g.2xlarge (8 core and 32G RAM). Tested 4 threads with 500_000 ops each. The results are slightly different because I didn't use the same configuration as you.. but overall, there's a noticeable improvement. Congrats for the work.
```
CPU Peak (%): Main: 33.4%, New: 32.8%, Difference: -0.6%
CPU Avg (%): Main: 30.3%, New: 19.8%, Difference: -34.65%
Malloc Avg Peak: Main: 2,508,459, New: 1,779,362, Difference: -29.05%
Heap Peak (MB): Main: 202 MB, New: 281 MB, Difference: +39.11%
Heap Avg (MB): Main: 132 MB, New: 128 MB, Difference: -3.03%
Heap Obj Avg: Main: 994,136, New: 970,217, Difference: -2.41%
```